# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACCESS Ready API - Adaptive ELL exit prep for K-12 students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get the Stripe subscription for a given owner (studentId for solo, teacherId for organization)
      parameters:
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
        - name: ownerType
          in: query
          required: true
          schema:
            type: string
            enum: [solo, organization]
      responses:
        "200":
          description: Subscription, if one exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: No subscription found
  /billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [billing]
      summary: Create a real Stripe Checkout Session for a plan; client redirects to the returned URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Stripe-hosted checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
...
  /billing/checkout-intent:
    post:
      operationId: createCheckoutIntent
      tags: [billing]
      summary: Create a Stripe subscription in incomplete state and return a client secret for an embedded (in-app) payment form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Client secret for confirming the subscription's initial payment with Stripe.js
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutIntent"
  /billing/portal:
    post:
      operationId: createBillingPortalSession
      tags: [billing]
      summary: Create a Stripe Billing Portal session for managing/canceling a subscription and payment methods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalBody"
      responses:
        "200":
          description: Stripe-hosted billing portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "404":
          description: No Stripe customer found for this owner
  /billing/admin/summary/{teacherId}:
    get:
      operationId: getBillingAdminSummary
      tags: [billing]
      summary: Seat count and estimated monthly cost for a teacher/district admin's roster
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
        - name: seatCount
          in: query
          required: false
          description: Override the seat count used for the estimated-cost calculation (e.g. a not-yet-submitted seat count the user is previewing before checkout). Defaults to the teacher's actual roster size.
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Seat and cost summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingAdminSummary"
  /billing/payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [billing]
      summary: List saved payment methods (cards) for an owner's Stripe customer
      security:
        - bearerAuth: []
      parameters:
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
        - name: ownerType
          in: query
          required: true
          schema:
            type: string
            enum: [solo, organization]
      responses:
        "200":
          description: Saved payment methods
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPaymentMethodsResponse"
  /billing/checkout-with-saved-card:
    post:
      operationId: checkoutWithSavedCard
      tags: [billing]
      summary: >
        Create and immediately confirm a new subscription using a previously
        saved payment method (off-session, no Elements form required).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutWithSavedCardBody"
      responses:
        "200":
          description: Subscription created and payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutWithSavedCardResponse"
...
components:
  schemas:
...
    Subscription:
      type: object
      description: Sourced live from stripe.subscriptions for the owner's Stripe customer.
      properties:
        id:
          type: string
        ownerId:
          type: string
        ownerType:
          type: string
          enum: [solo, organization]
        planId:
          type: string
          enum: [solo, organization]
        status:
          type: string
          enum: [active, trialing, past_due, canceled, incomplete, incomplete_expired, unpaid, paused]
        seatCount:
          type: integer
        currentPeriodEnd:
          type: string
          nullable: true
        cancelAtPeriodEnd:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - ownerId
        - ownerType
        - planId
        - status
        - seatCount
        - cancelAtPeriodEnd
        - createdAt
    CheckoutBody:
      type: object
      properties:
        ownerId:
          type: string
        ownerType:
          type: string
          enum: [solo, organization]
        planId:
          type: string
          enum: [solo, organization]
        seatCount:
          type: integer
          default: 1
      required:
        - ownerId
        - ownerType
        - planId
    PortalBody:
      type: object
      properties:
        ownerId:
          type: string
        ownerType:
          type: string
          enum: [solo, organization]
      required:
        - ownerId
        - ownerType
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    CheckoutIntent:
      type: object
      properties:
        clientSecret:
          type: string
          description: PaymentIntent client secret for confirming the subscription's initial payment with an embedded Stripe Elements form.
        subscriptionId:
          type: string
      required:
        - clientSecret
        - subscriptionId
    SavedPaymentMethod:
      type: object
      properties:
        id:
          type: string
          description: Stripe PaymentMethod ID (pm_…)
        brand:
          type: string
          description: Card brand, e.g. "visa", "mastercard"
        last4:
          type: string
        expMonth:
          type: integer
        expYear:
          type: integer
        isDefault:
          type: boolean
          description: True if this is the customer's default payment method
      required:
        - id
        - brand
        - last4
        - expMonth
        - expYear
        - isDefault
    SavedPaymentMethodsResponse:
      type: object
      properties:
        paymentMethods:
          type: array
          items:
            $ref: "#/components/schemas/SavedPaymentMethod"
      required:
        - paymentMethods
    CheckoutWithSavedCardResponse:
      type: object
      properties:
        subscriptionId:
          type: string
        status:
          type: string
      required:
        - subscriptionId
        - status
    CheckoutWithSavedCardBody:
      type: object
      properties:
        ownerId:
          type: string
        ownerType:
          type: string
          enum: [solo, organization]
        planId:
          type: string
          enum: [solo, organization]
        seatCount:
          type: integer
          minimum: 1
        paymentMethodId:
          type: string
          description: Stripe PaymentMethod ID to charge off-session
      required:
        - ownerId
        - ownerType
        - planId
        - paymentMethodId
    BillingAdminSummary:
      type: object
      properties:
        teacherId:
          type: string
        seatCount:
          type: integer
        pricePerSeatCents:
          type: integer
        estimatedMonthlyCostCents:
          type: integer
        currency:
          type: string
        displayPricePerSeat:
          type: string
          description: Server-formatted price-per-seat string ready to render as-is, e.g. "$20 / seat / month". Frontend must never derive this from pricePerSeatCents.
        displayEstimatedMonthlyCost:
          type: string
          description: Server-formatted estimated monthly cost string ready to render as-is, e.g. "$60 / month", already computed for the given (or requested) seat count.
        subscription:
          $ref: "#/components/schemas/Subscription"
      required:
        - teacherId
        - seatCount
        - pricePerSeatCents
        - estimatedMonthlyCostCents
        - currency
        - displayPricePerSeat
        - displayEstimatedMonthlyCost
...